+++ packages/doctor/src/getCustomESLintConfig.ts
@@ -1,12 +1,12 @@
 import * as fs from 'fs-extra';
-import * as path from 'path';
+import { join } from 'path';
 import { parse } from '@babel/parser';
 import traverse from '@babel/traverse';
 
 export default function getCustomESLintConfig(directory: string) {
 
   let config = {};
-  const configFilePath = path.join(directory, '.eslintrc.js');
+  const configFilePath = join(directory, '.eslintrc.js');
 
   if (fs.existsSync(configFilePath)) {
     const source = fs.readFileSync(configFilePath, { encoding: 'utf-8' });
@@ -21,6 +21,7 @@ export default function getCustomESLintConfig(directory: string) {
         if (node.callee.name === 'getESLintConfig' && node.arguments && node.arguments[1]) {
           const configNode = node.arguments[1];
           const configSource = source.substring(configNode.start, configNode.end);
+          // eslint-disable-next-line no-eval
           config = eval(`(${configSource})`);
         }
       },
